move utilities in fixes
